SurveyForge-main/code/requirement.txt

torch==2.1.0
transformers==4.44.2
sentence-transformers==2.7.0
accelerate==0.33.0
safetensors==0.4.4
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-community==0.3.0
langchain-core==0.3.12
langchain-huggingface==0.1.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.136
llama-index==0.11.9
llama-index-core==0.11.9
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.4
llama-index-llms-openai==0.2.5
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.5
openai==1.52.0
anthropic==0.34.2
faiss-cpu==1.9.0
faiss-gpu==1.7.2
pandas==2.2.2
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.1
pypdf==4.3.1
beautifulsoup4==4.12.3
openpyxl==3.1.5
striprtf==0.0.26
nltk==3.9.1
tiktoken==0.7.0
tokenizers==0.19.1
regex==2024.7.24
sqlalchemy==2.0.32
tinydb==4.8.0
requests==2.32.3
httpx==0.27.0
aiohttp==3.10.5
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.5.2
pyyaml==6.0.2
tqdm==4.66.4
click==8.1.7
h5py==3.10.0